+++ server/migrations/sql/20250911002435_create_settings_table.sql
@@ -16,7 +16,7 @@ CREATE TABLE IF NOT EXISTS settings
     show_daily_report               BOOLEAN NOT NULL DEFAULT true,
     show_fund_updates               BOOLEAN NOT NULL DEFAULT true,
     calculate_collection            INT NOT NULL DEFAULT 0,
-    total_amount                    INT NOT NUll DEFAULT 0,
+    total_amount                    FLOAT NOT NUll DEFAULT 0,
     updated_at                      TIMESTAMP    DEFAULT current_timestamp,
     updated_by                      VARCHAR(100) NOT NULL DEFAULT ''
 );
